:adhesive_bandage: fix sonarlint major issues

+++ src/core/app/centerWindow.cpp
@@ -14,8 +14,8 @@ void App::centerWindow(void)
 {
     sf::VideoMode desktop = sf::VideoMode::getDesktopMode();
     sf::Vector2i winPos(
-        desktop.width / 2.f - _width / 2.f,
-        desktop.height / 2.f - _height / 2.f);
+        (int)(desktop.width / 2.f - _width / 2.f),
+        (int)(desktop.height / 2.f - _height / 2.f));
 
     _window.setPosition(sf::Vector2i(
         min<int>(desktop.width - _width, winPos.x),
